Fix small bug wtih transform
When you create a gameObj that is a child of another game object, it took the current transform and substract that with the parent transform, which is expected but a different kind of behavior is need when load.
Now when load it will add the transform of parent of previously wasn't owned by other game object.